Change Count to float

+++ src/Models/Check.cs
@@ -8,6 +8,6 @@
         public Product Product { get; set; }
         public DateTime Date { get; set; }        
         public float Price { get; set; }
-        public int Count { get; set; }
+        public float Count { get; set; }
     }
 }
